Added support for log level
