+++ packages/perfrunner-core/src/utils/index.ts
@@ -1,5 +1,3 @@
 import { orderByAscending } from './array';
-import { debug, log, throwException } from './log';
 
 export { orderByAscending };
-export { debug, log, throwException };
